管理画面ダッシュボード及びマスタ

+++ app/laravel/database/migrations/pilates/2026_04_24_130145_create_locations_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->boolean('is_bookable')->default(true);
             $table->unsignedInteger('price_addon_per_session')->default(0);
             $table->string('address');
-            $table->decimal('base_fee', 8, 2)->nullable();
+            $table->unsignedInteger('base_fee')->nullable();
             $table->string('map_url')->nullable();
             $table->boolean('is_active')->default(true);
             $table->timestamps();
